Delete backup-RBC4-Braiding-Sweetgrass.yml
If it becomes too large (like RBC3 did), then there will be a memory error in the backup. To fix, you would have to turn off EDN backups.
My suggestion is to create 2 more repos (a private and a public) for the book club backup, to keep it separate from your personal backups. That way you can turn off EDN just for the book club, while keeping EDN backup it for the others.